end-devices/requirements.txt
psutil
requests
pika
python-dotenv
